Create eval_spoint_angles.m
eval_spoint_angles: Evaluates the angle between each zero and pole of a
transfer function and a given point in the s-plane. Output in either
radians or degrees, depending on the mode input. With a single output,
returns only the required controller contribution.

INPUTS:
- tf: Must be either a 'tf' or 'zpk', symbolic not implemented.
- mode: Must be either 'radians' or 'degrees'.
- s: Point on the s-plane to which the angles will be evaluated.

OUTPUTS:
- angles_zeros: Angular contribution for each of the tf zeros.
- angles_poles: Angular contribution for each of the tf poles.
- angles_controller: Angular contribution required from a cascade
controller in order for s-point to be in root locus of the resulting
transfer function.